refactor: centralize product picker modal
